Remove duplicated code and move it into one function Also fixed small bug with trimming text
