+++ src/connect_rule_window_purge_click.rs
@@ -4,6 +4,7 @@ use gtk::prelude::*;
 
 pub fn connect_rule_window_purge_click(gui_data: &GuiData) {
     let window_rules = gui_data.window_rules.clone();
+
     let radio_button_purge_both = gui_data.window_rules.purge.radio_button_purge_both.clone();
     let radio_button_purge_name = gui_data.window_rules.purge.radio_button_purge_name.clone();
     let radio_button_purge_extension = gui_data.window_rules.purge.radio_button_purge_extension.clone();
